pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "ibis-dsql"
version = "0.1.0"
description = "DSQL SQL compiler extension for Ibis"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "ibis-framework==11.0.0",
  "sqlglot==28.1.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.2,<9",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ibis_dsql"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
